requirements.txt
python-chess
fastapi
uvicorn
